+++ agents-manage-api/src/data/tools.ts
@@ -1,4 +1,4 @@
-import { managementServer } from '../index.js';
+import { managementServer } from '../server.js';
 import { getLogger } from '../logger.js';
 import dbClient from './db/dbClient.js';
