+++ agents-manage-api/src/routes/oauth.ts
@@ -19,7 +19,7 @@ import {
   getCredentialReference,
   updateCredentialReference,
 } from '@inkeep/agents-core';
-import { managementServer } from '../index.js';
+import { managementServer } from '../server.js';
 import { getLogger } from '../logger.js';
 import { oauthService, retrievePKCEVerifier } from '../utils/oauth-service.js';
 import dbClient from '../data/db/dbClient.js';
