+++ agents-run-api/package.json
@@ -7,12 +7,11 @@
   "type": "module",
   "license": "SEE LICENSE IN LICENSE.md",
   "scripts": {
-    "dev": "nodemon --watch src --ext ts --exec 'tsx src/index.ts'",
-    "dev:apis": "nodemon --watch src --ext ts --exec 'tsx src/index.ts'",
-    "dev:with-bypass": "PORT=3003 nodemon --watch src --ext ts --exec 'tsx src/index.ts'",
-    "dev:without-bypass": "PORT=3004 INKEEP_AGENTS_RUN_BYPASS_SECRET= nodemon --watch src --ext ts --exec 'tsx src/index.ts'",
+    "dev": "vite",
+    "dev:with-bypass": "PORT=3003 vite",
+    "dev:without-bypass": "PORT=3004 INKEEP_AGENTS_RUN_BYPASS_SECRET= vite",
     "build": "tsc",
-    "start": "node dist/index.js",
+    "start": "node dist/server.js",
     "test": "ENVIRONMENT=test vitest --run",
     "test:watch": "ENVIRONMENT=test vitest",
     "test:coverage": "ENVIRONMENT=test vitest --run --coverage",
@@ -24,8 +23,8 @@
     "@ai-sdk/react": "2.0.11",
     "@hono/node-server": "^1.14.3",
     "@hono/otel": "^0.4.0",
-    "@hono/zod-openapi": "^1.0.2",
     "@hono/swagger-ui": "^0.5.1",
+    "@hono/zod-openapi": "^1.0.2",
     "@inkeep/agents-core": "workspace:*",
     "@modelcontextprotocol/sdk": "^1.17.2",
     "@opentelemetry/api": "^1.9.0",
@@ -57,6 +56,7 @@
   },
   "devDependencies": {
     "@biomejs/biome": "2.1.4",
+    "@hono/vite-dev-server": "^0.20.1",
     "@types/ajv": "^1.0.4",
     "@types/jmespath": "^0.15.2",
     "@types/node": "^20.11.24",
@@ -65,6 +65,8 @@
     "nodemon": "^3.1.0",
     "tsx": "^4.7.1",
     "typescript": "^5.3.3",
+    "vite": "^7.1.4",
+    "vite-tsconfig-paths": "^5.1.4",
     "vitest": "^3.1.4"
   },
   "packageManager": "pnpm@10.11.0",
