+++ packages/agents-core/clean-package.config.json
@@ -1,0 +1,16 @@
+{
+    "replace": {
+        "main": "dist/index.js",
+        "exports": {
+            ".": "./dist/index.js",
+            "./schema": "./dist/schema/index.js",
+            "./types": "./dist/types/index.js",
+            "./validation": "./dist/validation/index.js",
+            "./client-exports": {
+                "types": "./dist/client-exports.d.ts",
+                "import": "./dist/client-exports.js"
+            },
+            "./package.json": "./package.json"
+        }
+    }
+}
