+++ packages/agents-core/package.json
@@ -4,19 +4,16 @@
   "description": "Core database schema, types, and validation schemas for Inkeep Agent Framework",
   "type": "module",
   "license": "SEE LICENSE IN LICENSE.md",
+  "main": "./src/index.ts",
   "exports": {
-    ".": "./dist/index.js",
-    "./schema": "./dist/schema/index.js",
-    "./types": "./dist/types/index.js",
-    "./validation": "./dist/validation/index.js",
-    "./client-exports": {
-      "types": "./dist/client-exports.d.ts",
-      "import": "./dist/client-exports.js"
-    }
+    ".": "./src/index.ts",
+    "./schema": "./src/db/schema.ts",
+    "./types": "./src/types/index.ts",
+    "./validation": "./src/validation/index.ts",
+    "./client-exports": "./src/client-exports.ts"
   },
   "scripts": {
     "build": "tsc",
-    "dev": "tsc --watch",
     "test": "vitest --run",
     "test:unit": "vitest --run src/__tests__ --exclude src/__tests__/integration/**",
     "test:integration": "vitest --run src/__tests__/integration/",
@@ -31,9 +28,18 @@
     "db:clean": "tsx src/db/clean.ts",
     "db:studio": "drizzle-kit studio",
     "db:check": "drizzle-kit check",
-    "db:reset-schema": "rm -rf drizzle/* && echo 'All migration files removed, generating new schema' && drizzle-kit generate"
+    "db:reset-schema": "rm -rf drizzle/* && echo 'All migration files removed, generating new schema' && drizzle-kit generate",
+    "prepack": "clean-package",
+    "postpack": "clean-package restore"
   },
+  "clean-package": "./clean-package.config.json",
   "dependencies": {
+    "@hono/node-server": "^1.14.3",
+    "@hono/zod-openapi": "^1.0.2",
+    "@libsql/client": "^0.15.7",
+    "@modelcontextprotocol/sdk": "^1.17.2",
+    "@nangohq/node": "^0.66.0",
+    "@nangohq/types": "^0.66.0",
     "@opentelemetry/api": "^1.9.0",
     "@opentelemetry/auto-instrumentations-node": "^0.62.0",
     "@opentelemetry/baggage-span-processor": "^0.4.0",
@@ -43,32 +49,27 @@
     "@opentelemetry/sdk-node": "^0.203.0",
     "@opentelemetry/sdk-trace-node": "^2.0.1",
     "@opentelemetry/semantic-conventions": "^1.34.0",
-    "@hono/zod-openapi": "^1.0.2",
-    "@libsql/client": "^0.15.7",
-    "@modelcontextprotocol/sdk": "^1.17.2",
-    "@hono/node-server": "^1.14.3",
-    "@nangohq/node": "^0.66.0",
-    "@nangohq/types": "^0.66.0",
     "ai": "5.0.11",
     "ajv": "^8.17.1",
     "ajv-formats": "^3.0.1",
     "dotenv": "^17.2.1",
+    "drizzle-orm": "^0.44.4",
+    "drizzle-zod": "^0.8.2",
     "exit-hook": "^4.0.0",
-    "ts-pattern": "^5.7.1",
     "hono": "^4.8.10",
     "jmespath": "^0.16.0",
     "keytar": "^7.9.0",
-    "drizzle-orm": "^0.44.4",
-    "drizzle-zod": "^0.8.2",
     "nanoid": "^5.0.9",
+    "ts-pattern": "^5.7.1",
     "zod": "^4.1.5"
   },
   "devDependencies": {
     "@biomejs/biome": "2.1.4",
     "@types/jmespath": "^0.15.2",
     "@types/node": "^20.11.24",
-    "drizzle-kit": "^0.31.4",
     "@vitest/coverage-v8": "^2.0.0",
+    "clean-package": "^2.2.0",
+    "drizzle-kit": "^0.31.4",
     "typescript": "^5.9.2",
     "vitest": "^3.1.4"
   },
