+++ packages/agents-core/tsconfig.json
@@ -13,7 +13,16 @@
     "declarationMap": true,
     "sourceMap": true,
     "allowSyntheticDefaultImports": true,
-    "resolveJsonModule": true
+    "resolveJsonModule": true,
+    "baseUrl": ".",
+    "paths": {
+      "@inkeep/agents-core": [
+        "./src/index.ts"
+      ],
+      "@inkeep/agents-core/*": [
+        "./src/*"
+      ]
+    }
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist", "**/*.test.ts"]
